refactor Module
